build: make API URL an environment variable
Make the API URL used by the extension change based
on whether the environment is development or production.
This is accomplished during the build process; when
we specify a dev build, Webpack will inject the
dev API URL into the emitted code, and the same thing
for the production build.