fix: fix prod sourcemaps not loading correctly
Add production sourcemap files to web accessible
resources in manifest.json so they are loaded
correctly. The sourcemap option we've chosen for
production creates separate sourcemap files
(non-inline), so they have to be specifically
whitelisted. Previously they were not whitelisted
which resulted in them not being loaded at all,
indicated by a net::ERR_BLOCKED_BY_CLIENT in the
console.

Note that the development sourcemaps never had this
problem since the sourcemap option we've chosen for
development makes them part of the bundled
code itself (inline), which is already allowed
as either content or background scripts.