refactor: remove manuscript caching
Remove caching of uploaded manuscripts in
localStorage. It was originally intended to
be a performance optimization: manuscripts
already uploaded to the server won't need
to be uploaded again. However, this assumes
that manuscript data doesn't change on the
dashboard after upload, which isn't an
assumption we can make currently. Additionally,
there is negligible performance gain; we send
all manuscripts in one HTTP request,
manuscripts are very small, and the max number
of manuscripts sent is 50. The difference in
number of HTTP requests and payload size
between caching or no caching isn't big enough
to cause any performance problems.
In conclusion, it is being removed because it
solves no problems, and introduces potential ones.