feat: configure monitoring for dev build
Configure development build to report errors on Sentry. This is done
so that we have a place to test monitoring functionality changes before
putting it into production. This build points to a 'dev' environment on
Sentry (achieved by the 'environment' option in Sentry.init), which
ensures that dev errors don't mix with actual production errors.
- Modify development webpack config to upload sourcemaps every reload if
  process.env.UPLOAD_SOURCEMAPS==true, to ensure accurate error tracing
  on Sentry. Uploading is optional because it takes several seconds,
  which wastes time if devs are testing non-monitoring functionality.

- Modify existing dev start command and add new dev start command
  to opt into/out of sourcemap upload on reload by setting the
  aforementioned environment variable.
  - Install new package cross-env to allow setting of environment
    variable on command line on any OS.