chore: simplify example .env files being tracked
Simplify the example .env files tracked in the repository down to just
2 files demonstrating what the public and private variables are for all
environments. The original intent of having dedicated example files
for each environment was to streamline setup and synchronization
of environment variables, but it scales badly as the number of
environment grows (n*2 files to maintain) and doesn't actually
solve the problem.