build: Auto-rebuild on webpack config change
Modifying any file that may change the build result should trigger a
rebuild automatically. Webpack config files fall into this category
since they are directly responsible for specifying what is outputted
to the final build. Therefore, changing a webpack config should trigger
a rebuild. However, currently it does not, which forces the developer
to manually restart the dev server to apply changes to the config.
Webpack does not provide a built-in solution to do this but a common
workaround is to use nodemon and configure it to restart the webpack dev
server every time the config file changes. This commit does exactly
that.