ci: make build zip easier to load in browser
Tweak release workflow so that when zipping the extension build that
is attached to the release on GitHub, the path to manifest.json
is no longer zipped_dir/build/manifest.json, but instead
zipped_dir/manifest.json. Because (Chromium) browsers expect the
manifest to be in the root of the directory specified to load the
extension from, this new dir structure makes figuring out which
directory to load simple, since devs can just specify the top-level
directory and not go down another level to top-level-dir/build.
See #92 for more context.